refactor: post slice -> tag slice
