+++ udong-frontend/app/domain/store/__test__/PostSlice.test.ts
@@ -31,7 +31,6 @@ const fakePost: PostState = {
     selectedPost: fakeListItemPost1,
     feedPosts: [],
     clubPosts: [fakeBoardPost1, fakeBoardPost2],
-    createPostTags: [],
 }
 
 jest.mock('next/config', () => () => ({
