+++ udong-frontend/app/domain/store/post/PostSelector.ts
@@ -29,16 +29,10 @@ const createPostError = createSelector(
     state => state.createPostError,
 )
 
-const createPostTags = createSelector(
-    postState,
-    state => state.createPostTags,
-)
-
 export const postSelector = {
     selectedPost,
     feedPosts,
     clubPosts,
     createdPostId,
     createPostError,
-    createPostTags,
 }
