+++ udong-frontend/app/ui/feature/club/board/__test__/BoardContainer.test.tsx
@@ -8,7 +8,6 @@ import { BoardContainer } from '../BoardContainer'
 const stubInitialState: PostState = {
     feedPosts: [],
     clubPosts: [],
-    createPostTags: [],
 }
 
 const mockStore = configureStore({
