+++ udong-frontend/app/ui/feature/club/post/detail/__test__/PostDetailContainer.test.tsx
@@ -29,7 +29,6 @@ const stubPostInitialState: PostState = {
     selectedPost: dummyPost,
     feedPosts: [],
     clubPosts: [],
-    createPostTags: [],
 }
 
 const stubCommentInitialState: CommentState = {
