+++ udong-frontend/app/ui/feature/club/post/scheduling/__test__/SchedulingCloseContainer.test.tsx
@@ -61,7 +61,6 @@ const postInitialState: PostState = {
     },
     feedPosts: [],
     clubPosts: [],
-    createPostTags: [],
 }
 
 export const mockStore = configureStore({
