+++ udong-frontend/app/ui/feature/club/post/upsert/create/__test__/PostCreateContainer.test.tsx
@@ -19,7 +19,6 @@ jest.mock('../../PostAdditionalFieldsView', () => ({
 const stubPostInitialState: PostState = {
     feedPosts: [],
     clubPosts: [],
-    createPostTags: [],
 }
 
 const mockStore = configureStore({
