+++ udong-frontend/app/ui/feature/club/tag/__test__/TagContainer.test.tsx
@@ -14,6 +14,7 @@ const stubInitialState: TagState = {
         updatedAt: '',
         users: [],
     }],
+    createPostTags: [],
 }
 
 const stubUserInitialState: UserState = {
