+++ udong-frontend/app/ui/feature/home/__test__/HomeContainer.test.tsx
@@ -17,7 +17,6 @@ const stubInitialState: ClubState = {
 const stubPostInitialState: PostState = {
     feedPosts: [],
     clubPosts: [],
-    createPostTags: [],
 }
 
 const mockStore = configureStore({
