+++ udong-frontend/app/ui/feature/home/feed/__test__/FeedContainer.test.tsx
@@ -8,7 +8,6 @@ import { FeedContainer } from '../FeedContainer'
 const stubPostInitialState: PostState = {
     feedPosts: [],
     clubPosts: [],
-    createPostTags: [],
 }
 
 const mockStore = configureStore({
